+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Improved the `xsql` errors mapping to `driver.ErrBadConn` 
+* Extended `retry.DoTx` test for to achieve equivalence with `retry.Retry` behaviour
+
 ## v3.33.0
 * Added `retry.DoTx` helper for retrying `database/sql` transactions 
 * Implemented `database/sql` driver over `ydb-go-sdk`